backend/chat-service/server/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Giang Ta <thoanggiang98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.13.4"
fastapi = "^0.63.0"
python-socketio = "^5.0.4"
websockets = "^8.1"
asyncpg = "^0.22.0"
tortoise-orm = "^0.16.21"
pytest = "^6.2.2"
requests = "^2.25.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.2.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
